finish traits, start transformations
