+++ rust/src/traits/samplers/bernoulli/mod.rs
@@ -137,25 +137,4 @@ pub fn sample_bernoulli_rational(prob: RBig, trials: Option<usize>) -> Fallible<
 }
 
 #[cfg(test)]
-mod test {
-    use super::*;
-    use crate::traits::samplers::test::*;
-
-    #[test]
-    fn test_bernoulli() {
-        [0.2, 0.5, 0.7, 0.9].iter().for_each(|p| {
-            let sampler = || {
-                if sample_bernoulli_float(*p, false).unwrap() {
-                    1.
-                } else {
-                    0.
-                }
-            };
-            assert!(
-                test_proportion_parameters(sampler, *p, 0.00001, *p / 100.),
-                "empirical evaluation of the bernoulli({:?}) distribution failed",
-                p
-            )
-        })
-    }
-}
+mod test;
